product model fixed, product showing, and UI fixes

+++ lib/app/modules/single_shop/transaction_and_refund/presentation/pages/exchange_sell_page.dart
@@ -1551,7 +1551,7 @@ List<Widget> getTabBarViews(
                                       Product product = Product(
                                           name: controller
                                               .customProductName.value,
-                                          sellingPrice: double.parse(controller
+                                          sellingPrice: int.parse(controller
                                               .customProductPrice.value),
                                           stock: 1);
